camshot.py
"""
camshot.py
-------------

Description:
    This module is used to test connection with camera over USB

Author:
    Umair Cheema <cheemzgpt@gmail.com>

Version:
    1.0.0

License:
    Apache License 2.0 (https://www.apache.org/licenses/LICENSE-2.0)

Date Created:
    2024-12-12

Last Modified:
    2024-12-13

Python Version:
    3.8+

Usage:
    Can be run from the command line /Terminal /Shell
    Example:
        python camshot.py

Dependencies:

"""
import cv2

cv2.namedWindow("preview")
vc = cv2.VideoCapture(0)

if vc.isOpened(): # try to get the first frame
    rval, frame = vc.read()
else:
    rval = False

while rval:
    cv2.imshow("preview", frame)
    rval, frame = vc.read()
    key = cv2.waitKey(20)
    if key == 27: # exit on ESC
        break

cv2.destroyWindow("preview")
vc.release()
