requirements.txt
ansitable==0.11.4
anyio==4.7.0
appnope==0.1.4
argcomplete==3.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
bleak==0.22.3
catkin-pkg==1.0.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colored==2.2.4
coloredlogs==15.0.1
comm==0.2.2
coverage==7.6.9
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docutils==0.21.2
empy==4.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.10.0
h11==0.14.0
hidapi==0.14.0.post4
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
identify==2.6.3
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.5.1
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
onnxslim==0.1.44
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pgraph-python==0.6.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
progress==1.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-repeat==0.9.3
pytest-rerunfailures==15.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qpsolvers==4.4.0
quadprog==0.1.13
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboticstoolbox-python==1.1.1
rpds-py==0.22.3
rtb-data==1.0.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
spatialgeometry==1.1.0
spatialmath-python==1.1.13
stack-data==0.6.3
-e git+https://github.com/jhavl/swift.git@efa25fb570aceed6070658b24bfe7526b8644fc8#egg=swift_sim
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.49
ultralytics-thop==2.0.13
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
xarm==0.0.4
